requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
sqlmodel==0.0.21
pydantic==2.9.2
python-dotenv==1.0.1

# Auth
passlib==1.7.4
python-jose[cryptography]==3.3.0

# OpenAI sentiment
openai==1.52.2
email-validator==2.2.0

